Broke down saml module for google apps
