+++ cas-server-support-saml/src/main/java/org/jasig/cas/support/saml/SamlServletContextListener.java
@@ -19,35 +19,18 @@
 
 package org.jasig.cas.support.saml;
 
-import org.apache.commons.lang3.StringUtils;
-import org.jasig.cas.support.saml.authentication.principal.GoogleAccountsServiceFactory;
 import org.jasig.cas.support.saml.authentication.principal.SamlService;
 import org.jasig.cas.support.saml.authentication.principal.SamlServiceFactory;
 import org.jasig.cas.support.saml.web.SamlValidateController;
-import org.jasig.cas.util.PrivateKeyFactoryBean;
-import org.jasig.cas.util.PublicKeyFactoryBean;
 import org.jasig.cas.util.UniqueTicketIdGenerator;
 import org.jasig.cas.web.AbstractServletContextInitializer;
 import org.jasig.cas.web.support.WebUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowire;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.FileSystemResource;
 import org.springframework.stereotype.Component;
-import org.springframework.util.ResourceUtils;
 
 import javax.servlet.ServletContextEvent;
 import javax.servlet.annotation.WebListener;
-import java.security.PrivateKey;
-import java.security.PublicKey;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Initializes the CAS root servlet context to make sure
@@ -62,10 +45,6 @@ public class SamlServletContextListener extends AbstractServletContextInitialize
     @Autowired
     @Qualifier("samlServiceFactory")
     private SamlServiceFactory samlServiceFactory;
-
-    @Autowired
-    @Qualifier("googleAccountsServiceFactory")
-    private GoogleAccountsServiceFactory googleAccountsServiceFactory;
 
     @Autowired
     @Qualifier("samlServiceTicketUniqueIdGenerator")
@@ -85,7 +64,6 @@ public class SamlServletContextListener extends AbstractServletContextInitialize
     @Override
     protected void initializeRootApplicationContext() {
         addServiceFactory(samlServiceFactory);
-        addServiceFactory(googleAccountsServiceFactory);
         addServiceTicketUniqueIdGenerator(SamlService.class.getCanonicalName(),
                 this.samlServiceTicketUniqueIdGenerator);
     }
@@ -94,96 +72,5 @@ public class SamlServletContextListener extends AbstractServletContextInitialize
     protected void initializeServletApplicationContext() {
         addControllerToCasServletHandlerMapping(SamlProtocolConstants.ENDPOINT_SAML_VALIDATE, samlValidateController);
     }
-
-    @Component
-    @Configuration
-    private static class GoogleAppsConfigurationInitializer {
-
-        private final Logger logger = LoggerFactory.getLogger(getClass());
-
-        @Value("${cas.saml.googleapps.publickey.file:}")
-        private String publicKeyLocation;
-
-        @Value("${cas.saml.googleapps.privatekey.file:}")
-        private String privateKeyLocation;
-
-
-        @Value("${cas.saml.googleapps.key.alg:}")
-        private String keyAlgorithm;
-
-        protected GoogleAppsConfigurationInitializer() {}
-
-        @Bean(name="googleAppsPrivateKey", autowire = Autowire.BY_NAME)
-        public PrivateKey getGoogleAppsPrivateKey() throws Exception {
-            if (!isValidConfiguration()) {
-                logger.debug("Google Apps private key bean will not be created, because it's not configured");
-                return null;
-
-            }
-            final PrivateKeyFactoryBean bean = new PrivateKeyFactoryBean();
-
-            if (this.privateKeyLocation.startsWith(ResourceUtils.CLASSPATH_URL_PREFIX)) {
-                bean.setLocation(new ClassPathResource(
-                    StringUtils.removeStart(this.privateKeyLocation, ResourceUtils.CLASSPATH_URL_PREFIX)));
-            } else if (this.privateKeyLocation.startsWith(ResourceUtils.FILE_URL_PREFIX)) {
-                bean.setLocation(new FileSystemResource(
-                        StringUtils.removeStart(this.privateKeyLocation, ResourceUtils.FILE_URL_PREFIX)));
-            } else {
-                bean.setLocation(new FileSystemResource(this.privateKeyLocation));
-            }
-
-            bean.setAlgorithm(this.keyAlgorithm);
-            logger.debug("Loading Google Apps private key from {} with key algorithm {}",
-                    bean.getLocation(), bean.getAlgorithm());
-
-            bean.afterPropertiesSet();
-
-
-            logger.debug("Creating Google Apps private key instance via {}", this.publicKeyLocation);
-            return bean.getObject();
-        }
-
-        @Bean(name="googleAppsPublicKey", autowire = Autowire.BY_NAME)
-        public PublicKey getGoogleAppsPublicKey() throws Exception {
-            if (!isValidConfiguration()) {
-                logger.debug("Google Apps public key bean will not be created, because it's not configured");
-                return null;
-            }
-
-            final PublicKeyFactoryBean bean = new PublicKeyFactoryBean();
-            if (this.publicKeyLocation.startsWith(ResourceUtils.CLASSPATH_URL_PREFIX)) {
-                bean.setLocation(new ClassPathResource(
-                        StringUtils.removeStart(this.publicKeyLocation, ResourceUtils.CLASSPATH_URL_PREFIX)));
-            } else if (this.publicKeyLocation.startsWith(ResourceUtils.FILE_URL_PREFIX)) {
-                bean.setLocation(new FileSystemResource(
-                        StringUtils.removeStart(this.publicKeyLocation, ResourceUtils.FILE_URL_PREFIX)));
-            } else {
-                bean.setLocation(new FileSystemResource(this.publicKeyLocation));
-            }
-
-            bean.setAlgorithm(this.keyAlgorithm);
-
-            logger.debug("Loading Google Apps public key from {} with key algorithm {}",
-                    bean.getResource(), bean.getAlgorithm());
-
-            bean.afterPropertiesSet();
-
-            logger.debug("Creating Google Apps public key instance via {}", this.publicKeyLocation);
-            return bean.getObject();
-        }
-
-        @Bean(name="serviceFactoryList")
-        public List getServiceFactoryList() {
-            return new ArrayList();
-        }
-
-        private boolean isValidConfiguration() {
-            return StringUtils.isNotBlank(this.privateKeyLocation)
-                    || StringUtils.isNotBlank(this.publicKeyLocation)
-                    || StringUtils.isNotBlank(this.keyAlgorithm);
-        }
-    }
-    
-
 
 }
