+++ cas-server-support-saml/src/test/java/org/jasig/cas/support/saml/AllTestsSuite.java
@@ -21,7 +21,6 @@ package org.jasig.cas.support.saml;
 
 import org.jasig.cas.support.saml.authentication.SamlAuthenticationMetaDataPopulatorTests;
 import org.jasig.cas.support.saml.authentication.SamlAuthenticationRequestTests;
-import org.jasig.cas.support.saml.authentication.principal.GoogleAccountsServiceTests;
 import org.jasig.cas.support.saml.authentication.principal.SamlServiceTests;
 import org.jasig.cas.support.saml.util.SamlCompliantUniqueTicketIdGeneratorTests;
 import org.jasig.cas.support.saml.web.flow.mdui.SamlMetadataUIParserActionTests;
@@ -38,7 +37,6 @@ import org.junit.runners.Suite;
  */
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
-        GoogleAccountsServiceTests.class,
         SamlServiceTests.class,
         SamlAuthenticationMetaDataPopulatorTests.class,
         SamlAuthenticationRequestTests.class,
