+++ cas-server-webapp/pom.xml
@@ -79,6 +79,12 @@
       <groupId>org.apache.logging.log4j</groupId>
       <artifactId>log4j-web</artifactId>
     </dependency>
+
+    <dependency>
+      <groupId>org.jasig.cas</groupId>
+      <artifactId>cas-server-support-saml-googleapps</artifactId>
+      <version>4.2.0-SNAPSHOT</version>
+    </dependency>
   </dependencies>
 
   <build>
