+++ cas-server-webapp/src/main/webapp/WEB-INF/cas.properties
@@ -357,9 +357,9 @@ tgc.signing.key=szxK-5_eJjs-aUj-64MpUZ-GPPzGLhYPLGl0wrYjYNVAGva2P0lLe6UGKGM7k8dW
 ##
 # Google Apps public/private key
 #
-# cas.saml.googleapps.publickey.file=classpath:DSAPrivateKey01.key
-# cas.saml.googleapps.privatekey.file=classpath:DSAPrivateKey01.key
-# cas.saml.googleapps.key.alg=DSA
+# cas.saml.googleapps.publickey.file=file:/etc/cas/public.key
+# cas.saml.googleapps.privatekey.file=file:/etc/cas/private.p8
+# cas.saml.googleapps.key.alg=RSA
 
 ##
 # WS-FED
